[TASK] Move clear-cache hook in `CacheService`
